Include lots of forgotten math.h and roundf() is included in C99.
git-svn-id: svn+ssh://svn.gna.org/svn/warzone/trunk@3586 4a71c877-e1ca-e34f-864e-861f7616d084

+++ lib/framework/trig.c
@@ -29,6 +29,9 @@
 
 #include <assert.h>
 #include <stdlib.h>
+
+#include <math.h>
+
 #include "types.h"
 #include "debug.h"
 #include "trig.h"
@@ -52,7 +55,7 @@ static float aSqrt[SQRT_ACCURACY];
 /* Initialise the Trig tables */
 BOOL trigInitialise(void)
 {
-	float val = 0.0f, inc = 2.0f * (float)M_PI / TRIG_DEGREES;
+	float val = 0.0f, inc = 2.0f * M_PI / TRIG_DEGREES;
 	int i;
 
 	// Initialise the tables
